+++ packages/orq-rc/RELEASES.md
@@ -7779,3 +7779,13 @@ Based on:
 - [typescript v4.7.0-rc.21] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.21 - packages/orq-rc
+
+## 2026-03-26 10:55:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.3 (2.869.25) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.22] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.22 - packages/orq-rc